+++ client/src/layouts/admin/programs/index.js
@@ -21,6 +21,7 @@ import Tooltip from "@mui/material/Tooltip";
 import Dialog from "@mui/material/Dialog";
 import DialogContent from "@mui/material/DialogContent";
 import * as programService from "services/program";
+import Swal from "sweetalert2";
 import ProgramForm from "./forms";
 
 const columns = [
@@ -50,11 +51,29 @@ function Programs() {
 
   const handleDeleteProgram = async (programId) => {
     try {
-      await programService.deleteProgram(programId);
-      setPrograms(programs.filter((program) => program.programId !== programId));
+      Swal.fire({
+        title: "Are you sure?",
+        text: "Do you really want to delete this record? This process cannot be undone.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#FF0000",
+        confirmButtonText: "Delete",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          programService.deleteProgram(programId);
+          setPrograms(programs.filter((program) => program.programId !== programId));
+          Swal.fire("Deleted", "Record has been deleted.", "success");
+        } else if (result.isDenied) {
+          Swal.fire("Deletion has been cancelled", "", "info");
+        }
+      });
     } catch (error) {
       if (error.response && error.response.status === 404) {
-        alert("Program may have already been deleted");
+        Swal.fire({
+          icon: "error",
+          title: "Error",
+          text: "This record may have already been deleted.",
+        });
       }
     }
   };
@@ -109,7 +128,7 @@ function Programs() {
                 </IconButton>
                 <Dialog open={open} onClose={handleClose} fullWidth>
                   <DialogContent>
-                    <ProgramForm />
+                    <ProgramForm onClose={handleClose} />
                   </DialogContent>
                 </Dialog>
               </MDBox>
